feat(core): Increase metadata offsets to 64-bit for large datasets
The previous on-disk dataset format used 32-bit integers for metadata offsets, which limited the maximum size of a generated dataset file to 4GB.

This commit updates the `Meta` struct and the corresponding serialization/deserialization logic in `CoreWorkload` and `Dataset` to use 64-bit integers (`uint64_t`). This change removes the 4GB file size limitation, enabling the generation and use of much larger workloads.

As part of this change, the default record and operation count in the `workload_kv_session` file has been increased from 1M to 5M to test this new capability.

BREAKING CHANGE: The on-disk dataset format has been modified. Dataset files generated with previous versions of the tool are no longer compatible.

+++ core/core_workload.cc
@@ -507,12 +507,12 @@ void CoreWorkload::PrepareOpsForFile(std::ofstream& ofs, int n, bool is_loading)
 
     Meta meta;
     meta.key_offset = ofs.tellp();
-    meta.key_offset += sizeof(char) + sizeof(uint32_t) + sizeof(Meta);
+    meta.key_offset += sizeof(char) + sizeof(uint64_t) + sizeof(Meta);
     meta.op_specific_data_offset = meta.key_offset + sizeof(uint32_t) + key_len;
     meta.next_record_offset = meta.op_specific_data_offset + op_specific_data_len;
-
+    
     ofs.put(static_cast<char>(item.type));
-    uint32_t meta_len = sizeof(Meta);
+    uint64_t meta_len = sizeof(Meta);
     ofs.write(reinterpret_cast<const char*>(&meta_len), sizeof(meta_len));
     ofs.write(reinterpret_cast<const char*>(&meta), sizeof(meta));
